chore: add temporary page and refactor auth folder
